moved settings to utils
