+++ mrmustard/__init__.py
@@ -18,7 +18,7 @@
 from rich import print
 
 from ._version import __version__
-from .settings import *
+from .utils.settings import *
 
 
 def version():
