+++ mrmustard/utils/settings.py
